Remove the use of SocketAddr as it slowed parsing for no benefit.
